requirements.txt
numpy
llama-cpp-python
fastcore
spacy
nltk
flask
fire
pyzmq
nbdev